arm64: DT: trinket: Declare quirks on MPM and MPM-GPIO intcs
These interrupt controllers are managed a little different from
the other legacy platforms and, in order for these to peacefully
live together, we need to declare quirks somewhere.

I chose to declare quirks to SM6125 as a shorter way of "getting
there", instead of modifying any and every legacy platform to
include "functionality" that we have lost with the new firmwares
in this new SoCs.